Modify TranslationsImporter
